chore: update to librelane 2.4.2, cleanup
