+++ Contributing.md
@@ -15,7 +15,7 @@ Ensure that your local git clone also has an `upstream` remote. You can add it v
 ```sh
 git remote get-url upstream
 > fatal: No such remote 'upstream'
-git remote add upstream https://github.com/Cloud-V/DFFRAM`
+git remote add upstream https://github.com/AUCOHL/DFFRAM`
 ```
 
 You can start a rebase by typing:
@@ -39,17 +39,7 @@ git push -fu origin main
 Then, create your pull request. If all went well, the CI should pass.
 
 ## Note to maintainers
+
 If the number of commits is low, rebase and merge, but if it is high (5 commits or more), squash and merge.
 
 Squash and merge makes for a cleaner commit history, but will require pretty much everyone to rebase any work they've had off the main branch with a great degree more difficulty.
-
-# Legal
-By submitting a pull request, you (the Contributor) hereby grant The American University in Cairo a world-wide, royalty-free, non-exclusive license under intellectual property rights (other than patent or trademark) licensable by such Contributor to:
-
-* use, reproduce, make available, modify, display, perform, distribute, and otherwise exploit its Contributions, either on an unmodified basis, with Modifications, or as part of a Larger Work;
-
-* under Patent Claims of such Contributor to make, use, sell, offer for sale, have made, import, and otherwise transfer either its Contributions or its Contributor Version.
-
-Copyright notice: above text adapted from the Mozilla Public License, version 2.0.
-
-You are entitled and encouraged to add your name and email to the AUTHORS document at the root of the repository under 'Other Contributors', as well as include your name in any modified or created file.
