+++ models/rf/model.v
@@ -1,24 +1,8 @@
-/*
-    Copyright ©2020-2022 The American University in Cairo
-
-    This file is part of the DFFRAM Memory Compiler.
-    See https://github.com/Cloud-V/DFFRAM for further info.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
+// SPDX-License-Identifier: Apache-2.0
+// Copyright ©2020-2022, The American University in Cairo
 /* 
-	DFFRFile
-    Mohamed Shalan (mshalan@aucegypt.edu)
+	Module: DFFRF_2R1W
+    Author: Mohamed Shalan (mshalan@aucegypt.edu)
 	32x32 Register File with 2RW1W ports and clock gating for SKY130A 
 	~ 3550 Cells
 	< 2ns (no input or output delays)
