+++ placeram/cli.py
@@ -1,20 +1,6 @@
 # -*- coding: utf8 -*-
-# Copyright ©2020-2022 The American University in Cairo
-#
-# This file is part of the DFFRAM Memory Compiler.
-# See https://github.com/Cloud-V/DFFRAM for further info.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# SPDX-License-Identifier: Apache-2.0
+# Copyright ©2020-2025, The American University in Cairo
 
 import os
 import re
@@ -23,6 +9,7 @@ import traceback
 try:
     import odb
     import utl
+    from openroad import Tech, Design
 except ImportError:
     print(
         """
@@ -63,7 +50,9 @@ class Placer:
         tap_distance,
     ):
         # Initialize Database
-        self.db = odb.dbDatabase.create()
+        self.ord_tech = Tech()
+        self.design = Design(self.ord_tech)
+        self.db = self.ord_tech.getDB()
 
         odb.read_db(self.db, odb_in)
