+++ placeram/placeable.py
@@ -1,3 +1,6 @@
+# -*- coding: utf8 -*-
+# SPDX-License-Identifier: Apache-2.0
+# Copyright ©2020-2022, The American University in Cairo
 import os
 import re
 import sys
