+++ platforms/gf180mcuD/gf180mcu_fd_sc_mcu7t5v0/block_definitions.v
@@ -1,22 +1,5 @@
-/*
-    Copyright ©2020-2021 The American University in Cairo and the Cloud V Project.
-
-    This file is part of the DFFRAM Memory Compiler.
-    See https://github.com/Cloud-V/DFFRAM for further info.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
-
+// SPDX-License-Identifier: Apache-2.0
+// Copyright ©2020-2022, The American University in Cairo
 // Add 1x2 binary decoder
 `default_nettype none
