+++ requirements.txt
@@ -1,3 +1,3 @@
-openlane==2.0.0b16
+librelane>=2.4.0,<3
 pyyaml
 cloup
