--- shell.nix
@@ -1,23 +1,0 @@
-# Copyright 2023 Efabless Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-{
-    pkgs ? import <nixpkgs> {}
-}:
-
-let openlane-src = pkgs.fetchFromGitHub {
-    owner = "efabless";
-    repo = "openlane2";
-    rev = "83ec6c32add40006cc70d951745667d30193f51d";
-    sha256 = "sha256-9Xms6eRf3yyaFJVVjk+uPYYM+EDRl4GYAYYUK5QhiLc=";
-}; in import "${openlane-src}/shell.nix" {}
